# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List organizations
      responses:
        "200":
          description: Organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationCollection"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{organizationId}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organizations/{organizationId}/demo/reset:
    post:
      operationId: resetOrganizationDemoState
      tags: [demo]
      summary: Reset organization to deterministic demo state
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Demo state reset summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoResetResponse"
  /organizations/{organizationId}/teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams for an organization
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamCollection"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create team
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /organizations/{organizationId}/teams/{teamId}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TeamIdParam"
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update team details
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TeamIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: Team updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete team
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TeamIdParam"
      responses:
        "204":
          description: Team deleted
  /organizations/{organizationId}/teams/{teamId}/ownership:
    put:
      operationId: assignTeamOwnership
      tags: [ownership]
      summary: Assign team ownership
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TeamIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignOwnershipRequest"
      responses:
        "200":
          description: Ownership assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamOwnership"
  /organizations/{organizationId}/ownership/teams:
    get:
      operationId: listTeamOwnerships
      tags: [ownership]
      summary: List team ownership assignments
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Team ownership assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamOwnershipCollection"
  /organizations/{organizationId}/positions:
    get:
      operationId: listPositions
      tags: [positions]
      summary: List positions for an organization
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Positions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionCollection"
    post:
      operationId: createPosition
      tags: [positions]
      summary: Create position
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePositionRequest"
      responses:
        "201":
          description: Position created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /organizations/{organizationId}/positions/{positionId}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get position
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PositionIdParam"
      responses:
        "200":
          description: Position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
    patch:
      operationId: updatePosition
      tags: [positions]
      summary: Update position details
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PositionIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePositionRequest"
      responses:
        "200":
          description: Position updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
    delete:
      operationId: deletePosition
      tags: [positions]
      summary: Delete position
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PositionIdParam"
      responses:
        "204":
          description: Position deleted
  /organizations/{organizationId}/positions/{positionId}/ownership:
    put:
      operationId: assignPositionOwnership
      tags: [ownership]
      summary: Assign position ownership
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PositionIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignOwnershipRequest"
      responses:
        "200":
          description: Ownership assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionOwnership"
  /organizations/{organizationId}/ownership/positions:
    get:
      operationId: listPositionOwnerships
      tags: [ownership]
      summary: List position ownership assignments
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Position ownership assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionOwnershipCollection"
  /organizations/{organizationId}/people:
    get:
      operationId: listPeople
      tags: [people]
      summary: List people for an organization
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: People
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeopleCollection"
    post:
      operationId: createPerson
      tags: [people]
      summary: Create person
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonRequest"
      responses:
        "201":
          description: Person created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
  /organizations/{organizationId}/people/{personId}:
    get:
      operationId: getPerson
      tags: [people]
      summary: Get person
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PersonIdParam"
      responses:
        "200":
          description: Person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
    patch:
      operationId: updatePerson
      tags: [people]
      summary: Update person
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PersonIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonRequest"
      responses:
        "200":
          description: Person updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
    delete:
      operationId: deletePerson
      tags: [people]
      summary: Delete person
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PersonIdParam"
      responses:
        "204":
          description: Person deleted
  /organizations/{organizationId}/actions:
    get:
      operationId: listActions
      tags: [actions]
      summary: List actions for an organization
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Actions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionCollection"
    post:
      operationId: createAction
      tags: [actions]
      summary: Create action
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActionRequest"
      responses:
        "201":
          description: Action created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
  /organizations/{organizationId}/positions/{positionId}/actions:
    get:
      operationId: listPositionActions
      tags: [actions]
      summary: List actions for a position context
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PositionIdParam"
      responses:
        "200":
          description: Position actions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionCollection"
  /organizations/{organizationId}/positions/{positionId}/execution-summary:
    get:
      operationId: getPositionExecutionSummary
      tags: [actions]
      summary: Get execution summary for a position
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PositionIdParam"
      responses:
        "200":
          description: Position execution summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionExecutionSummary"
  /organizations/{organizationId}/people/{personId}/actions:
    get:
      operationId: listPersonActions
      tags: [actions]
      summary: List actions for a person context
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PersonIdParam"
      responses:
        "200":
          description: Person actions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionCollection"
  /organizations/{organizationId}/actions/{actionId}:
    get:
      operationId: getAction
      tags: [actions]
      summary: Get action
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/ActionIdParam"
      responses:
        "200":
          description: Action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
    delete:
      operationId: deleteAction
      tags: [actions]
      summary: Delete action
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/ActionIdParam"
      responses:
        "204":
          description: Action deleted
...
components:
  parameters:
    OrganizationIdParam:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TeamIdParam:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PositionIdParam:
      name: positionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    DemoResetResponse:
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
        teams:
          type: integer
        positions:
          type: integer
        people:
          type: integer
        actions:
          type: integer
        policies:
          type: integer
      required: [organizationId, teams, positions, people, actions, policies]
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, createdAt, updatedAt]
    CreateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
      required: [name, slug]
    OrganizationCollection:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
      required: [items]
...
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
        parentTeamId:
          type: string
          format: uuid
      required: [name]
...
    TeamCollection:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Team"
      required: [items]
...